requirements.txt
Django>=5.0,<6.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
celery>=5.3.0
redis>=5.0.0
python-dotenv>=1.0.0
whitenoise>=6.6.0
django-crispy-forms>=2.1
crispy-tailwind>=1.0.0
django-allauth>=0.61.0
django-filter>=23.5
Pillow>=10.0.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
gunicorn>=21.0.0
